document-rag/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.6
boto3==1.26.137
wheel>=0.38.0
setuptools>=65.0.0
pymongo==4.3.3
python-docx==0.8.11
PyPDF2==3.0.1
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
azure-identity>=1.12.0
azure-ai-textanalytics>=5.3.0
faiss-cpu>=1.7.4
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
tokenizers>=0.13.0
langchain-openai>=0.0.5
langchain-community>=0.0.10 